Check that graph is a bipartite weighted graph.

A bipartite weighted graph is a bipartite graph with edge weights.

isBipartiteWeightedGraph(graph, n1, n2)

Arguments

graph

object check.

n1

size of the first list of records.

n2

size of the second list of records.

Value

TRUE if graph is a bipartite weighted graph, FALSE otherwise.

Examples

graph <- igraph::graph.bipartite(types=c(rep(TRUE, 3), rep(FALSE, 2)), edges = c(1, 4, 3, 5), directed = FALSE)
igraph::E(graph)$weight = c(1.5, 0.5)
n1 <- 3
n2 <- 2
isBipartiteWeightedGraph(graph, n1, n2)
#> [1] TRUE
# Returns: TRUE

graph[]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>                       
#> [1,] .   . .   1.5 .  
#> [2,] .   . .   .   .  
#> [3,] .   . .   .   0.5
#> [4,] 1.5 . .   .   .  
#> [5,] .   . 0.5 .   .  
# 5 x 5 sparse Matrix of class "dgCMatrix"
# 
# [1,] .   . .   1.5 .  
# [2,] .   . .   .   .  
# [3,] .   . .   .   0.5
# [4,] 1.5 . .   .   .  
# [5,] .   . 0.5 .   .